+++ pom.xml
@@ -24,7 +24,7 @@
 		<!-- dataunits -->
 		<module>uv-dataunit-files</module>
 		<module>uv-dataunit-rdf</module>
-		<module>uv-dataunit-tabular</module>
+		<module>uv-dataunit-relational</module>
 		<!-- dataunit helpers and support libraries -->
 		<module>uv-dataunit-helpers</module>
 	</modules>
