+++ uv-dataunit-helpers/src/main/java/eu/unifiedviews/helpers/dataunit/relationalhelper/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Helpers providing easier way to work with {@link eu.unifiedviews.dataunit.relational.RelationalDataUnit}
+ */
+package eu.unifiedviews.helpers.dataunit.relationalhelper;
