--- uv-dataunit-helpers/src/main/java/eu/unifiedviews/helpers/dataunit/tabularhelper/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Helpers providing easier way to work with {@link eu.unifiedviews.dataunit.tabular.TabularDataUnit}
- */
-package eu.unifiedviews.helpers.dataunit.tabularhelper;
