+++ uv-dataunit-relational/src/main/java/eu/unifiedviews/dataunit/relational/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Definition of {@link eu.unifiedviews.dataunit.relational.RelationalDataUnit} which holds DB tables as enhanced entries inside {@link eu.unifiedviews.dataunit.MetadataDataUnit}.
+ */
+package eu.unifiedviews.dataunit.relational;
