--- uv-dataunit-tabular/src/main/java/eu/unifiedviews/dataunit/tabular/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Definition of {@link eu.unifiedviews.dataunit.tabular.TabularDataUnit} which holds DB tables as enhanced entries inside {@link eu.unifiedviews.dataunit.MetadataDataUnit}.
- */
-package eu.unifiedviews.dataunit.tabular;
